Fixes throttle issues introduced by 3.4
Summary:
Throttle range used to be 0-1000 but is now a floating 0-1.
GCS_Control was structured to be 0-1000.
Now is 0-1.
Stabilize No Pilot check for pilot stick position removed.

Test Plan: tested in sim and real world

Reviewers: tony

Reviewed By: tony

Differential Revision: http://git.vergeaero.com/D132

+++ ArduCopter/GCS_Mavlink.cpp
@@ -1691,9 +1691,9 @@ void GCS_MAVLINK_Copter::handleMessage(mavlink_message_t* msg)
 
         case MAV_CMD_GCS_CONTROL_THROTTLE: {
             if(packet.param2 == GCS_THROTTLE_MODE_TRIM){
-                copter.gcs_control.set_throttle_trim((int16_t)packet.param1);
+                copter.gcs_control.set_throttle_trim(packet.param1);
             }else {
-                copter.gcs_control.set_throttle((int16_t)packet.param1);
+                copter.gcs_control.set_throttle(packet.param1);
             }
 
             copter.gcs_control.set_throttle_mode((int8_t)packet.param2);
